+++ tests/testthat/_problems/test-length_range-162.R
@@ -1,5 +1,0 @@
-# Extracted from test-length_range.R:162
-
-# test -------------------------------------------------------------------------
-manyStates <- c(1, 2, 4, 8, 16, 32, 64, 128, 256, 512)
-expect_silent(result <- MaximumLength.numeric(manyStates))